requirements.txt
pandas
FreeSimpleGUI
pulp
openpyxl
matplotlib
pyyaml
xlsxwriter
PyYaml